fix: removed useDevice
