+++ .deco/blocks/pages-Ecoplus-730038.json
@@ -20,8 +20,29 @@
               "desktop": "https://data.decoassets.com/content-hub/096f588e-623a-42eb-92da-6c076997671c/principal_11zon.webp",
               "mobile": "https://data.decoassets.com/content-hub/fce93308-fcae-40bc-9ba8-1ff8cb3cc972/principal-mobile_11zon.webp"
             }
+          },
+          {
+            "loadingOptions": {
+              "loading": "eager",
+              "fetchPriority": "high",
+              "preload": true
+            },
+            "src": {
+              "desktop": "https://data.decoassets.com/content-hub/096f588e-623a-42eb-92da-6c076997671c/principal_11zon.webp",
+              "mobile": "https://data.decoassets.com/content-hub/fce93308-fcae-40bc-9ba8-1ff8cb3cc972/principal-mobile_11zon.webp"
+            }
           }
-        ]
+        ],
+        "configs": {
+          "pagination": {
+            "enabledMobile": true,
+            "enabledDesktop": true
+          },
+          "navigation": {
+            "enabledDesktop": true,
+            "enabledMobile": true
+          }
+        }
       }
     },
     {
@@ -1004,7 +1025,8 @@
         "configs": {
           "pagination": {
             "enabledMobile": true,
-            "clickable": true
+            "clickable": true,
+            "enabledDesktop": false
           },
           "lazy": true,
           "navigation": {
