+++ sections/Lps/BannerMediaSection.tsx
@@ -9,7 +9,7 @@ import type { IResponsiveImage } from "site/types/ResponsiveImage.d.ts";
 import type { ISection } from "site/types/Section.d.ts";
 import type { ISliderConfigs } from "site/types/Slider.d.ts";
 import { DefaultBannerMedia } from "site/configs/BannerMediaSection.ts";
-import { useDevice } from "@deco/deco/hooks";
+import { useIsDesktop } from "site/hooks/useIsDesktop.ts";
 
 /**
  * @title {{#id}}{{id}}{{/id}}{{^id}}Ícone{{/id}}
@@ -56,7 +56,7 @@ export default function BannerMediaSliderSection({
   icons = DefaultBannerMedia.icons
 }: Props) {
   const id = useId();
-  const device = useDevice();
+  const isDesktop = useIsDesktop();
 
   if (!banners?.length) return null;
 
@@ -78,10 +78,10 @@ export default function BannerMediaSliderSection({
     spaceBetween,
     centeredSlides,
     pagination: {
-      enabled: device === "desktop" ? pagination?.enabledDesktop : pagination?.enabledMobile
+      enabled: isDesktop.value ? pagination?.enabledDesktop : pagination?.enabledMobile
     },
     navigation: {
-      enabled: device === "desktop" ? navigation?.enabledDesktop : navigation?.enabledMobile
+      enabled: isDesktop.value ? navigation?.enabledDesktop : navigation?.enabledMobile
     },
     breakpoints: {
       768: {
