+++ sections/Lps/BannerSliderSection.tsx
@@ -5,6 +5,7 @@ import type { IResponsiveImage } from "site/types/ResponsiveImage.d.ts";
 import type { ISection } from "site/types/Section.d.ts";
 import type { ISliderConfigs } from "site/types/Slider.d.ts";
 import { DefaultBannerSection } from "site/configs/BannerSliderSection.ts";
+import { useIsDesktop } from "site/hooks/useIsDesktop.ts";
 
 /**
  * @description Seção com um slider de banners.
@@ -32,8 +33,33 @@ export default function BannerSliderSection({
   banners = DefaultBannerSection.banners,
   configs
 }: Props) {
-  const id = useId();
-  
+  const id = useId(); 
+  const isDesktop = useIsDesktop();
+
+  const { pagination, navigation, slidesPerViewResponsive } = configs ?? {};
+
+  const sliderConfig = {
+    ...configs,
+    slidesPerView: slidesPerViewResponsive?.mobile ?? 1,
+    pagination: {
+      enabled: isDesktop.value ? pagination?.enabledDesktop : pagination?.enabledMobile,
+    },
+    navigation: {
+      enabled: isDesktop.value ? navigation?.enabledDesktop : navigation?.enabledMobile,
+    },
+    breakpoints: {
+      768: {
+        slidesPerView: slidesPerViewResponsive?.tablet ?? 1
+      },
+      1024: {
+        slidesPerView: slidesPerViewResponsive?.desktop ?? 1,
+        pagination: {
+          enabled: pagination?.enabledDesktop
+        }
+      }
+    }
+  } as ISliderConfigs;
+
   if (!banners?.length) return null;
 
   const defaultPropsBanners = banners.map(banner => {
@@ -56,7 +82,7 @@ export default function BannerSliderSection({
       classesContainer="banner-slider-section"
     >
       <BannerSlider
-        configs={configs}
+        configs={sliderConfig}
         rootId={id}
         banners={defaultPropsBanners}
       />
