+++ static/tailwind.css
@@ -2802,9 +2802,6 @@ input.tab:checked + .tab-content,
 .cursor-pointer {
   cursor: pointer;
 }
-.resize {
-  resize: both;
-}
 .list-inside {
   list-style-position: inside;
 }
